feat: generate email verification token
After user registers on the app generate and store a verification token inside the VerificationToken model with fields such as email as the provided email, expires in 1hr and token generated from uuid. But when user tries to login without verifying the email, it will restrict the login and generate a new token after destroying the old token corresponding to the provided email id if present already.