feat: generate password reset token
Generate a password reset token and store that token in the database. For doing this we need to first check if there is any existing token, this is done by the getResetTokenByEmail function which accepts the email of the user and returns the reset token. Using this function we check  if there is an existing token, if yes we then delete that token and create a new reset token in the database with email, expires and token values.