feat: new-password server action
The function accepts two params, form field values and a token retrieved from the password reset link on the email provided by the user. It checks for the token in database and then updates the password field of the user with the hashed password. In end it deletes the token from the resetPasswordToken model.