feat: login action includes 2FA verification
Implement a check to create a 2FA code and send the email of the same to the provided email during login. It first checks if 2FA is enabled, if yes then it checks for the input code field. If we have the code it then retrieves the 2FA token using the email and checks if that token is equal to the code provided or has expired. If not then it deletes that token from the db. It then creates the tow-factor confirmation if not already present else deletes the confirmation.